fix: strict clippy (-D warnings) across the FX effort's crates
CI's quality gate was failing on reverb-dsp; swept every crate this
effort touched — reverb-dsp, delay-dsp, eq-dsp, saturate-dsp,
tune-dsp, pitch-dsp, multiband-dsp, world-sys, signal-fx,
fts-modulation, pattern-ui, signal-macromod, and the
meter/pitch/unison/saturate/limiter plugin shells — all now pass
cargo clippy --all-targets -- -D warnings.

Real changes (not just allows): needless returns removed, iterator
forms where they read better, delay-dsp's modulation tests module
moved below the items it precedes, derived Default for MeterParams,
struct-literal test config in macromod, slice::from_ref in tune tests.
Allows only where the lint is wrong for the code: index loops spanning
parallel arrays + self borrows, the one-live-variant delay engine
enum, RT-safe large-Err channel submits.

+++ apps/plugins/saturate/src/lib.rs
@@ -175,7 +175,7 @@ impl Plugin for FtsSaturate {
         _context: &mut impl InitContext<Self>,
     ) -> bool {
         for p in &mut self.pre {
-            p.set_sample_rate(buffer_config.sample_rate as f32);
+            p.set_sample_rate(buffer_config.sample_rate);
             p.reset();
         }
         true
